State Machine Assignment
The player character has three states: Normal, Rage and Weak.
Normal is the initial state.
If the Normal character killed 3 enemies in the last 3 seconds, it goes to Rage, otherwise if it killed no enemies in the last 3 seconds, it goes to Weak.
Rage and Weak states last for 5 seconds.